[no-issue]: Extract getOrganisationUnitsByCountry
